backend/requirements.txt
django~=4.1.7
django-cors-headers~=3.14.0
djangorestframework~=3.14.0
djangorestframework-simplejwt~=5.2.2
django-recaptcha~=3.0.0
psycopg2-binary~=2.9.5
requests~=2.26.0
beautifulsoup4~=4.10.0
